+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.14.0] - 2019-08-20
+### Added
+- Add command `zeroconfig-package-scripts sync`
+
+## [3.13.4] - 2019-08-19
+### Fixed
+- `if (fs.pathExists(...))` to `if (fs.pathExistsSync(...))` in `startBrowser`
+
+## [3.13.3] - 2019-08-13
+### Added
+- Add `process.env.PUBLIC_URL`
+- Available use `REACT_APP_*` process.env in app. <https://create-react-app.dev/docs/adding-custom-environment-variables>
+- Available use env in *.html files (eg. `%PUBLIC_URL%`)
+
 ## [3.13.1] - 2019-08-12
 ### BREAKING CHANGES
 - Update `eslint` to `@6`
@@ -80,6 +94,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 - tsconfig's `strictFunctionTypes` to `true`
 
+[3.14.0]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.13.4...3.14.0
+[3.13.4]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.13.3...3.13.4
+[3.13.3]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.13.1...3.13.3
 [3.13.1]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.12.0...3.13.1
 [3.12.0]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.10.0...3.12.0
 [3.10.0]: https://github.com/react-zeroconfig/react-zeroconfig/compare/3.9.0...3.10.0